+++ images/win/scripts/Installers/Windows2019/Initialize-VM.ps1
@@ -8,7 +8,13 @@ function Disable-InternetExplorerESC {
     $UserKey = "HKLM:\SOFTWARE\Microsoft\Active Setup\Installed Components\{A509B1A8-37EF-4b3f-8CFC-4F3A74704073}"
     Set-ItemProperty -Path $AdminKey -Name "IsInstalled" -Value 0 -Force
     Set-ItemProperty -Path $UserKey -Name "IsInstalled" -Value 0 -Force
-    Stop-Process -Name Explorer -Force -ErrorAction Continue
+
+    $ieProcess = Get-Process -Name Explorer -ErrorAction SilentlyContinue
+
+    if ($ieProcess){
+        Stop-Process -Name Explorer -Force -ErrorAction Continue
+    }
+
     Write-Host "IE Enhanced Security Configuration (ESC) has been disabled."
 }
